refactor: Session을 독립적인 Aggregate Root로 분리, 그에 따른 패키지 구조 조정
- `Session`이 항상 `Course`나 `Curriculum`의 일부로 존재하는 것은 아니라는 점을 발견했습니다.
- 기존 설계에서는 `Session`이 `Course` 애그리것의 생명주기에 강하게 종속되어 있었습니다.
-  이로 인해 `Course`나 `Curriculum`에 속하지 않는 독립적인 세션을 모델링하는 데 제약이 있었습니다.
  - 예를 들어 한 커리큘럼/과정에 참여하는 특정 학생 4명에 대한 멘토링 세션을 만들고 싶을 때 기존 설계에서는 특정 커리큘럼/과정에 속한 멤버가 모두 해당 세션의 참여자로 취급될 수 있는 문제가 있습니다.

### 주요 변경 사항
- **도메인 분리**: `Session`을 `Course` 에서 분리하고, 자체 Repository를 갖도록 변경
- **패키지 구조 조정**: 도메인 경계를 명확히 하기 위해 `domain.session` 패키지 생성 및 관련 클래스 이동
- **느슨한 결합**: `Session`이 `Course`나 `Curriculum`을 ID로 참조하도록 변경하여 결합도 감소
