+++ src/main/java/me/chan99k/learningmanager/domain/member/CreateAccountRequest.java
@@ -1,4 +1,4 @@
-package me.chan99k.learningmanager.domain.auth;
+package me.chan99k.learningmanager.domain.member;
 
 public record CreateAccountRequest(
 	String email, String password, Long memberId
