+++ src/main/java/me/chan99k/learningmanager/domain/session/SessionType.java
@@ -1,0 +1,5 @@
+package me.chan99k.learningmanager.domain.session;
+
+public enum SessionType {
+	ONLINE, OFFLINE
+}
