--- src/main/java/me/chan99k/learningmanager/domain/study/SessionType.java
@@ -1,5 +1,0 @@
-package me.chan99k.learningmanager.domain.study;
-
-public enum SessionType {
-	ONLINE, OFFLINE
-}
